--- assets/mixins/imageMixin.js
@@ -1,7 +1,0 @@
-import { loadImages } from '../prismic/imageTools'
-
-export default {
-  mounted() {
-    this.$nextTick(loadImages)
-  },
-}
